Refactor Config component and add ConfigBoard component
