+++ app/components/ConfigBoard/ConfigBoard.module.css
@@ -1,0 +1,10 @@
+.configForm {
+  font-size: 20px;
+  display: grid;
+  grid-template-columns: repeat(6, 1fr);
+  flex-direction: row;
+  gap: 64px;
+  width: 100%;
+  margin-block: 1em;
+  align-items: center;
+}
